+++ services/core/shared/core/compat/sdk_v2/index.js
@@ -21,6 +21,8 @@ const recentBlocksCache = require('../../helpers/recentBlocksCache');
 const coreCache = require('./coreCache');
 const coreApi = require('./coreApi');
 
+const { request } = require('./request');
+
 const { getTotalNumberOfDelegates, getDelegateRankByUsername, reload } = require('./delegateCache');
 
 const {
@@ -30,11 +32,10 @@ const {
 	validateTimestamp,
 } = require('./epochTime');
 const { setCoreVersion, getCoreVersion, mapParams } = require('./coreVersionCompatibility');
-
 const { getConstants } = require('./constants');
+const { getBlocks } = require('./blocks');
 
 const {
-	getBlocks,
 	getNetworkStatus,
 	getNetworkConstants,
  } = coreApi;
@@ -295,8 +296,8 @@ const getReadyStatus = () => readyStatus;
 const nop = () => {};
 
 module.exports = {
-	get: coreApi.request,
-	request: coreApi.request,
+	get: request,
+	request,
 	parseAddress,
 	validateAddress,
 	validatePublicKey,
@@ -332,6 +333,7 @@ module.exports = {
 	peerStates,
 	setReadyStatus,
 	setCoreVersion,
+	getCoreVersion,
 	getReadyStatus,
 	getEpochUnixTime,
 	getUnixTime,
