+++ services/core/shared/core/index.js
@@ -62,6 +62,7 @@ const {
 	peerStates,
 	setReadyStatus,
 	setCoreVersion,
+	getCoreVersion,
 	getReadyStatus,
 	getUnixTime,
 	getTotalNumberOfDelegates,
@@ -105,6 +106,7 @@ module.exports = {
 	peerStates,
 	setReadyStatus,
 	setCoreVersion,
+	getCoreVersion,
 	getReadyStatus,
 	getUnixTime,
 	EMAcalc,
